+++ packages/api-extractor-model/src/model/ModelReferenceResolver.ts
@@ -6,6 +6,7 @@ import { type ApiItem, ApiItemKind } from '../items/ApiItem.js';
 import { ApiItemContainerMixin } from '../mixins/ApiItemContainerMixin.js';
 import { ApiParameterListMixin } from '../mixins/ApiParameterListMixin.js';
 import type { ApiEntryPoint } from './ApiEntryPoint.js';
+import type { ApiMethod } from './ApiMethod.js';
 import type { ApiModel } from './ApiModel.js';
 import type { ApiPackage } from './ApiPackage.js';
 
@@ -114,11 +115,21 @@ export class ModelReferenceResolver {
 			if (memberSelector === undefined) {
 				if (foundMembers.length > 1) {
 					const foundClass: ApiItem | undefined = foundMembers.find((member) => member.kind === ApiItemKind.Class);
+					const foundEvent: ApiItem | undefined = foundMembers.find((member) => member.kind === ApiItemKind.Event);
 					if (
 						foundClass &&
 						foundMembers.filter((member) => member.kind === ApiItemKind.Interface).length === foundMembers.length - 1
 					) {
 						currentItem = foundClass;
+					} else if (
+						foundMembers.every((member) => member.kind === ApiItemKind.Method && (member as ApiMethod).overloadIndex)
+					) {
+						currentItem = foundMembers.find((member) => (member as ApiMethod).overloadIndex === 1)!;
+					} else if (
+						foundEvent &&
+						foundMembers.filter((member) => member.kind === ApiItemKind.Method).length === foundMembers.length - 1
+					) {
+						currentItem = foundEvent;
 					} else {
 						result.errorMessage = `The member reference ${JSON.stringify(identifier)} was ambiguous`;
 						return result;