+++ packages/discord.js/src/util/APITypes.js
@@ -161,6 +161,11 @@
  */
 
 /**
+ * @external APIOverwrite
+ * @see {@link https://discord-api-types.dev/api/discord-api-types-v10/interface/APIOverwrite}
+ */
+
+/**
  * @external APIPartialEmoji
  * @see {@link https://discord-api-types.dev/api/discord-api-types-v10/interface/APIPartialEmoji}
  */
@@ -403,6 +408,11 @@
 /**
  * @external InteractionResponseType
  * @see {@link https://discord-api-types.dev/api/discord-api-types-v10/enum/InteractionResponseType}
+ */
+
+/**
+ * @external InviteType
+ * @see {@link https://discord-api-types.dev/api/discord-api-types-v10/enum/InviteType}
  */
 
 /**