+++ packages/scripts/src/generateSplitDocumentation.ts
@@ -240,7 +240,12 @@ function itemExcerptText(excerpt: Excerpt, apiPackage: ApiPackage) {
 				// dapi-types doesn't have routes for class members
 				// so we can assume this member is for an enum
 				if (meaning === 'member' && path && 'parent' in path) {
-					href += `/enum/${path.parent}#${path.component}`;
+					// unless it's a variable like FormattingPatterns.Role
+					if (path.parent.toString() === '__type') {
+						href += `#${token.text.split('.')[0]}`;
+					} else {
+						href += `/enum/${path.parent}#${path.component}`;
+					}
 				} else if (meaning === 'type' || meaning === 'var') {
 					href += `#${token.text}`;
 				} else {
@@ -253,26 +258,35 @@ function itemExcerptText(excerpt: Excerpt, apiPackage: ApiPackage) {
 				};
 			}
 
-			const resolved = token.canonicalReference
-				? resolveCanonicalReference(token.canonicalReference, apiPackage)
-				: null;
+			if (token.canonicalReference) {
+				const resolved = resolveCanonicalReference(token.canonicalReference, apiPackage);
 
-			if (!resolved) {
+				if (!resolved) {
+					return {
+						text: token.text,
+					};
+				}
+
+				const declarationReference = apiPackage
+					.getAssociatedModel()
+					?.resolveDeclarationReference(token.canonicalReference, apiPackage);
+				const foundItem = declarationReference?.resolvedApiItem ?? resolved.item;
+
 				return {
 					text: token.text,
+					resolvedItem: {
+						kind: foundItem.kind,
+						displayName: foundItem.displayName,
+						containerKey: foundItem.containerKey,
+						uri: resolveItemURI(foundItem),
+						packageName: resolved.package?.replace('@discordjs/', ''),
+						version: resolved.version,
+					},
 				};
 			}
 
 			return {
 				text: token.text,
-				resolvedItem: {
-					kind: resolved.item.kind,
-					displayName: resolved.item.displayName,
-					containerKey: resolved.item.containerKey,
-					uri: resolveItemURI(resolved.item),
-					packageName: resolved.package?.replace('@discordjs/', ''),
-					version: resolved.version,
-				},
 			};
 		}
 
@@ -328,7 +342,7 @@ function itemTsDoc(item: DocNode, apiItem: ApiItem) {
 					if (!foundItem && !resolved) {
 						return {
 							kind: DocNodeKind.LinkTag,
-							text: null,
+							text: codeDestination.memberReferences[0]?.memberIdentifier?.identifier ?? null,
 						};
 					}
 
@@ -951,11 +965,16 @@ export async function generateSplitDocumentation({
 
 			const members = entry.members
 				.filter((item) => {
-					if (item.kind !== 'Function') {
-						return true;
+					switch (item.kind) {
+						case ApiItemKind.Function:
+							return (item as ApiFunction).overloadIndex === 1;
+						case ApiItemKind.Interface:
+							return !entry.members.some(
+								(innerItem) => innerItem.kind === ApiItemKind.Class && innerItem.displayName === item.displayName,
+							);
+						default:
+							return true;
 					}
-
-					return (item as ApiFunction).overloadIndex === 1;
 				})
 				.map((item) => ({
 					kind: item.kind,